consistencize matches of patterns and queries
