+++ spheroscope/patterns.py
@@ -1,182 +1,18 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import os
 import re
 
 from flask import (Blueprint, current_app, jsonify, render_template, request,
                    session)
-from pandas import concat, read_csv
+from pandas import concat
 
 from .auth import login_required
 from .corpora import init_corpus, read_config
-from .database import Pattern, Query
-from .queries import patch_query_results
+from .database import Pattern, Query, get_patterns
+from .queries import run_queries, create_subcorpus, add_gold, evaluate, patch_query_results
 
 bp = Blueprint('patterns', __name__, url_prefix='/patterns')
-
-
-def run_queries(queries, cwb_id):
-    """collect matches of all queries belonging to one pattern as
-    dataframe.
-
-    index:
-    - <str> query_name: name of the query
-    - <int> match: match position of query
-    - <int> matchend: matchend position of query
-
-    columns:
-    - corpus_config['display']['p_show']: p-attribute realization sequences of context
-    - corpus_config['display']['s_show']: s-attribute realizations at match
-    - for each slot in query['anchors']['slots']:
-       - slot_START: start position of slot
-       - slot_END: end position of slot
-       for each p-att:
-         - slot_p-att: p-attribute realization sequence of slot
-
-    """
-
-    # init corpus
-    corpus_config = read_config(cwb_id)
-    corpus = init_corpus(corpus_config)
-
-    # run all queries belonging to a pattern
-    matches_list = list()
-    for query in queries:
-
-        current_app.logger.info("query: %s", query.name)
-        query = query.serialize()
-
-        # get dump
-        dump = corpus.query(
-            cqp_query=query['cqp'],
-            context=corpus_config['query']['context'],
-            context_break=corpus_config['query']['context_break'],
-            corrections=query['anchors']['corrections'],
-            match_strategy=corpus_config['query']['match_strategy']
-        )
-
-        # retreive concordance
-        matches = dump.concordance(
-            form='slots',
-            p_show=dict(corpus_config['display'])['p_show'],
-            s_show=dict(corpus_config['display'])['s_show'],
-            cut_off=None,
-            slots=query['anchors']['slots']
-        )
-
-        # add name
-        matches['query'] = query['meta']['name']
-
-        # translate anchors points in slot_START and slot_END for all slots
-        for slot in query['anchors']['slots']:
-
-            df = dump.df.reset_index()
-            df.index = dump.df.index
-
-            # select relevant columns
-            columns = query['anchors']['slots'][slot]
-            columns = [columns] if isinstance(columns, int) else columns
-            df = df[columns]
-
-            if df.shape[1] == 1:
-                # only one column: start and end are the same
-                df.columns = ["_".join([str(slot), 'START'])]
-                df["_".join([str(slot), 'END'])] = df["_".join([str(slot), 'START'])]
-            elif df.shape[1] == 2:
-                # two columns (start and end)
-                df.columns = ["_".join([str(slot), 'START']), "_".join([str(slot), 'END'])]
-
-            # join slots to matches dataframe
-            matches = matches.join(df)
-
-        # append to result list
-        matches_list.append(matches.reset_index())
-
-    # create output dataframe
-    result = concat(matches_list).set_index(['query', 'match', 'matchend'])
-
-    return result
-
-
-def add_gold(result, cwb_id, pattern, s_cwb='tweet_id', s_gold='tweet'):
-    """add gold annotation to result of run_queries()
-
-    """
-    try:
-        gold = read_csv(
-            os.path.join("library", cwb_id, "gold", "adjudicated.tsv"),
-            sep="\t", index_col=0
-        )
-    except FileNotFoundError:
-        result['TP'] = None
-    else:
-        # pre-process gold
-        gold = gold.loc[gold['pattern'] == pattern].rename(
-            {s_gold: s_cwb, 'annotation': 'TP'}, axis=1
-        )
-        # join explicit TPs and FPs
-        result = result.reset_index()
-        result = result.merge(gold[[s_cwb, "TP"]], on=s_cwb, how='left')
-        result = result.set_index(['query', 'match', 'matchend'])
-
-    return result
-
-
-def evaluate(tps):
-    """evaluate "TP" column of result of run_queries()
-
-    :param pd.Series tps: column with True / False annotations
-
-    """
-
-    # get TPs, FPs, precision
-    N = len(tps)
-    tps = tps.value_counts().to_dict()
-    tps['TP'] = tps.pop(True, 0)
-    tps['FP'] = tps.pop(False, 0)
-    tps['N'] = N
-
-    try:
-        tps['prec'] = tps['TP'] / (tps['FP'] + tps['TP'])
-    except ZeroDivisionError:
-        tps['prec'] = 'nan'
-
-    return tps
-
-
-def create_subcorpus(df, slot):
-    """transform one slot of result of run_queries() into a valid dump
-    (empty dataframe multi-indexed by match and matchend)
-
-    """
-
-    # select appropriate columns
-    column_start = "_".join([str(slot), "START"])
-    column_end = "_".join([str(slot), "END"])
-    dump = df.reset_index(drop=True)[[column_start, column_end]]
-
-    # only take rows where there's a match, sort ascending and deduplicate
-    dump = dump.fillna(-1, downcast='infer')
-    dump = dump[dump[column_start] != -1]
-    dump = dump.sort_values(by=column_start)
-    dump = dump.drop_duplicates()
-
-    # post-proc: behaviour when start > end: start = end
-    dump[column_end][
-        dump[column_start] > dump[column_end]
-    ] = dump[column_start][
-        dump[column_start] > dump[column_end]
-    ]
-
-    # rename, convert to int, set index
-    dump = dump.rename(
-        columns={column_start: 'match', column_end: 'matchend'}
-    ).set_index(
-        ['match', 'matchend']
-    )
-
-    return dump
 
 
 def hierarchical_query(p1, slot, p2):
@@ -275,14 +111,7 @@ def index():
 @bp.route('/api')
 @login_required
 def patterns():
-    patterns = Pattern.query.order_by(Pattern.id).all()
-    # FIXME this conversion should go when the new database is in
-    patterndict = [{
-        "id": abs(p.id),
-        "template": p.template,
-        "explanation": p.explanation,
-        "retired": p.id < 0
-    } for p in patterns]
+    patterndict = get_patterns()
     return jsonify(patterndict)
 
 
@@ -304,18 +133,27 @@ def pattern(id):
 def matches(id):
     """retrieve matches of all queries belonging to one pattern"""
 
+    # mapping of s-att that contains gold annotation
+    s_cwb = 'tweet_id'
+    s_gold = 'tweet'
+
+    # get matches
     cwb_id = session['corpus']['resources']['cwb_id']
     queries = Query.query.filter_by(pattern_id=id).order_by(Query.name).all()
     matches = run_queries(queries, cwb_id)
 
-    matches = add_gold(matches, cwb_id, id)
-    tps = evaluate(matches['TP'])
+    # add gold
+    matches = add_gold(matches, cwb_id, id, s_cwb, s_gold)
+    tps = evaluate(matches, s_cwb)
 
     # cut_off
     matches = matches.sample(int(request.args.get('cut_off', 100)))
 
+    # patch for frontend
+    matches = patch_query_results(matches)
+
     return render_template('queries/standalone_result_table.html',
-                           result=patch_query_results(matches),
+                           result=matches,
                            tps=tps)
